195113: Separate options to insert quotes on typing '=' / completion
This CL separates a combined option for adding quotes on typing '='
and attribute completion into two options. This is required to enable
automatic quotes insertion on attribute completion (thus fixing
http://b.android.com/195113) without bumping into non-obvious
behaviour documented in the issue
https://youtrack.jetbrains.com/issue/IDEA-113332

Old option corresponds to adding quotes on inserting '=' and is
already disabled in AndroidInitialConfigurator, the new option is
turned on by default.

Change-Id: I7b5853b36481823e210956d09700e8acf2576994