Implement authentication for the builtin webserver
Upon Android Studio start, it ensures there is a 'user.token'
file in the user's studio config dir (something like
 ~/.AndroidStudio20/config/user.token, only accessible by that user.
This file contains a cryptographic seed that is in text format.

The builtin webserver is modified to require basic authentication
in the form of user: '_token_', password: the contents of the 'user.token'
file above. Alternatively the token is allowed to be passed as a prefix
of the server path, see below why this is needed.

The BuiltInWebBrowserUrlProvider has been modified to launch
the urls with a prefix of the token in the path of the url.
This is needed as while most browsers support urls in the form of
'http://user:pass@host', Internet Explorer does not.

The SSH and HTTP git plugins have been modified to pass the right
username and password combination to the XmlRpc apis for Basic Auth.

Users who want to access the builtin webserver from the command-line
can run something like:

   curl -D -  -u _token_:`cat user.token` \
    http://localhost:63342/api/file?file=path-to-file

Change-Id: I9dfdccbd0616a26762705321364aa57e13db2ccc
Fixes: 124680708
Fixes: 137372765