Attempt to fix the missing dispose of editor.
Whenever a project is closed that has an open inspector we get
an error about a missing dispose of an editor. This is an attempt
to fix this problem.
The change is in a sensitive area and may not be 100% safe.

Change-Id: Ia92c3b838aa22a63199a37c01e0c719b66ea9c41
(cherry picked from commit f2b70e295810bcdb20fb56aab02e33a60feed6d5)