Fixes restarter flow on Linux and Mac.
On Mac it was using the name of patch for the distribution without
a jdk. This CL changes the property that tells studio that there is
a bundled jdk.

On linux the restarter path was incorrect not using the /system subfolder

The issue is still present on Windows.

Change-Id: I3de0710ee287003a9e5060fb71dc11b8467bcd48
(cherry picked from commit b029d85e440ca027fac19874d4e0c740ac172214)