a11y: Fix infinite recursion issue in ExpandedItemRendererComponent
The fix that was previously made did not work in all cases, and
we now have access to a general purpose AccessibleContext delegate
class that allows fixing the infinite parent chain issue in all
cases.

An example of the infinite recursion occurs when selecting items in
the "Ignored Schemas and DTDs" list of the "Languages & Frameworks >
Schemas and DTDs" settings page (and using calling the getObjectDepth
API, which can be done with Access Bridge Explorer).

Change-Id: I85dd7a8802d7945effeb886c4196046d09db75b7
(cherry picked from commit e4777bffba5831754b7266641f93e5826fc0838e)