Integrate analytics-library into Android Studio
ensures that analytics tracker & publisher are initialized,
relies on user to have opt-ed-in in Android Studio UI.

We want the Android Studio usage tracking to be available
from the platform, so this CL adds the entire metrics code
into the platform itself. The android plugin will depend on
the same libraries. We expect these libraries to be bundled
inside the android plugin in the IntelliJ distribution.

cherry-picked from 5332ffcb26217c1f0db762f214a9021ee0e8997d

Change-Id: I808e0333aae7e4a98734f4e74eb02f64ae7d5d9e