Move posting to analytics to a separate executor
Posting analytics (GA or google logs) is not batched. If some code
generates lots of events in a short time, each event becomes a new
runnable that is posted on the application pooled thread. There are two
problems with this:
 a) We could accidentally create too many threads, and the application
  thread pool is created as an unbounded queue.
 b) If there is any issue with the actual post (such as the JDK lock
  contention bug https://bugs.openjdk.java.net/browse/JDK-8133070),
  then each of those runnables could either take a long time or stall,
  causing even more pressure.

Studio 2.3 avoids this issue by adding proper batching of events. This
CL simply works around this in 2.2 by adding a separate executor that
limits the number of active threads and pending events, and dropping
requests that come in when the queue is full.

Fix for https://code.google.com/p/android/issues/detail?id=219563

Change-Id: Id3372d30cb941a5eb186d152e3c7e17d42e4fdaa