225130: Allow Studio to start up when running under a JRE
A bunch of debugging machinery indirectly references classes
from tools.jar (in particular com.sun.jdi.Value), and these
are aggressively opened at startup, which causes the IDE to
abort startup.

This CL (a) gracefully degrades in this case, and (b) adds
a system health monitor which warns about this problem.

Change-Id: I8b6e77a12d779ddc0883d2d043b8062dc6c7ac76