No Run/Debug configuration selected after loading a project
The root cause is a change made to IJ 2017.3
(https://github.com/JetBrains/intellij-community/commit/6312824af962657fec337f5f07263fa5e8f9c8a7)
that indirectly affected Android Studio, due to the way Android Studio
"inserts" the "app" run configuration during project load.

A fix for the underlying issue was made to IntelliJ (see
https://github.com/JetBrains/intellij-community/commit/3eba342e34a33bfc2169e449175bfd6b85998032)
but that fix is quite big and fixes another issue (run configuration
are not alphabetically sorted).

So, for Android Studio 3.1, we take a subset of the actual fix
(this change) that is less risky. Basically, we revert 2 lines
of initial code change that broke the behavior, that is we want
to ensure "projectRunConfigurationFirstLoaded" is called in all
cases.

See full details in bug 72204389 comments.

Change-Id: I22bd641c2391d36a53c3d66262b47bcbf5cb8820
Fixes: 72204389
Test: n/a