Prevent sealing the build view when sync is rerun from build view
Sync issues are not shown when sync is rerun using the refresh project
button on the build view, this is caused because we seal the build view
when the rerun action is performed.

This CL reuses the sealBuildViewOnSucess parameter for rerun action.

Bug: 74360539
Test: Manually tested.
Change-Id: Ief1b6c9c86f380777aa48d8755a8ccba0576a4bc
(cherry picked from commit d5a435c24dba27e844c0ce27231e0e9db5fd6b29)