Change EDT latency measurement to opt-in only.
IdeEventQueue contains instrumentation to measure the latency of IDE
events and is turned on by default. There is an ArrayIndexOutOfBounds
error that happens due to negative values passed to HdrHistogram from
this instrumentation.

This CL makes this instrumentation opt-in instead of opt-out, which
effectively disables it by default.

Bug: 110113356
Test: Verified that event latency measurements are disabled.
Change-Id: I72fa4328bc6a61c7d370f717cd3adff797ebb3b3
Exempt-From-Owner-Approval: owners don't have permission to vote on this branch