Fix repository URL to work with repository server
Up to AS 3.1, the URL sent to the plugins server would be of the form:

http://plugins.jetbrains.com/plugins/list/?uuid=3458baa2-9b96-4203-bbc5-d8104fb96a08&build=AI-xxx.yyy.zz

where "AI-xxx.yyy.zz" is the "API" build number of the product.

In 3.2 and later, we made a change to send URL of the form:

http://plugins.jetbrains.com/plugins/list/?uuid=3458baa2-9b96-4203-bbc5-d8104fb96a08&build=xxx.yyy.zz.aa.bb

where "xxx.yyy.zz.aa.bb" is the build number, and the last 2 numbers are the AS version (32 or 33) and
the AS build number.

The plugins server does not seem to understand this format and returns an empty list of plugins when
using the new URL.

This change reverts the behavior back to what it was in 3.1, but only for downloading plugins,
i.e. without changing the build number values of AS itself.

Bug: 111734668
Test: n/a
Change-Id: I2bf44caf4eeeb5143a8f63404f0b3885980c2e55
(cherry picked from commit 9b2cd51b21424f5446a9e4a4aa6281dbc17b96d2)