Initialize analytics differently when testing
Our platform code now initializes analytics settings very early in the
process. That initializer always assumes that $HOME is writable. This
means that tests have to explicitly initialize analytics with test
specific settings. This is not always possible, e.g. if there is a
plugin that is compatible with both IntelliJ and AS, that plugin's tests
cannot do any AS specific setup.

This CL initializes changes our setup code to initialize analytics
differently depending on whether we are running in test mode or not.

Bug: 117884305
Test: existing tests pass
Change-Id: I8659f77cf462c12434087421f80a7664366a12ab
(cherry picked from commit 3cce4ef45d3042b31a7fa99ff36bf520e5d97931)