ActionButtons background should include the entire button.
A change in IDEA restricted the background drawn for hover to only
include the icon as a side effect of that change.

Change this to draw the background for the entire button rectangle.
Also update the focus ring to match.
There is a bug filed here IDEA-199027 to track this.

Change-Id: I3ddb7c33192d669b27e3be9800fd91c45fcdc5f9
Fixes: 120773515
Test: NA
(cherry picked from commit ca92e662956077ec479e15962b5730e7db705716)