Don't include aswb-blaze plugin in Linux ASWB beta builds
We're planning to upload these builds to google3, where the IntelliJ
team will perform Beta releases of ASWB for us. The version of the
aswb-blaze plugin built here will be incompatible with the version
of g3plugins they build at release time, so we should just leave it to
them to drop in a compatible version of aswb-blaze.

We release ASWB on Mac directly from ADRT to dl.google.com and those
releases don't include g3plugins, so we still want ASWB on Mac beta
builds to include the aswb-blaze plugin.

This CL meets both requirements by moving the override of
copyAdditionalFiles that copies over the aswb-blaze plugin out of
AswbProperties and into AswbProperties.AswbMacDistributionCustomizer
instead.

Bug: 123402507
Test: test_studio.py
Change-Id: I40ccc749f2a7f5c8ab758b5e57fd3622a78877b7