Do not walk the tree to re-expand nodes on each change.
Instead expand nodes when the first child is being added to an auto-
expandable node in the Swing tree model.

For that make isAutoExpandableNode property read-only and do not try
to auto-expand errors this way. Instead delay auto-selection of the
first error properly so that it can always expand the tree up to the
error node.

Bug: 146593375
Test: n/a (manually tested)
Change-Id: I1e4c82b5227cc4cdc3585d36e17adbf5e547894c
(cherry picked from commit e05e840e665f6d59130aac803df2bbe018ac9e05)