Add entitlements.xml to macOS distribution zip
Adds a _codesign directory to root of macOS distribution zip that
contains files needed for signing.

macOS binaries need to be signed and notarized in order for them to work
on macOS Catalina and above. In order to notarize an application, all
executable binaries need to be signed with a hardened runtime, which
limits the privileges the binary has. In order to run the JVM, the
application needs a set of entitlements that are supplied during
signing. The _codesign directory will contain all of the files needed
for signing of studio.

This CL updates AndroidStudioProperties to include the default
entitlements used by IntelliJ in the codesign folder. In addition, the
the macOS distribution builder has been updated to put the _codesign
folder in the root of the zip archive, as the _codesign folder would
otherwise be placed inside Android Studio.app/Contents. It does this by
excluding the _codesign folder from the application bundle contents and
then copying the contents of the _codesign folder instead to a directory
also named _codesign at the root of the zip archive.

Test: N/A
Bug: 148305163
Change-Id: I376167135313f4a24e6f71c3e94ec879ce1cc033
(cherry picked from commit 3d13084c7ab72e61185bf8a66e7dea89bb5939b6)