Exclude dbus-java from macOS distribution
dbus is not needed by the macOS version of studio. The jar file also
contains a dylib that was built with an OSX SDK version lower than 10.9,
making it incompatible with Apple's new notarization requirements. In
order to be able to notarize studio, the jar must be removed from the
distribution.

Test: N/A
Bug: 148288696
Change-Id: Iae68782b5527c87f08cdd718048681ecf7f6b1ee
(cherry picked from commit d4fa5d38eda748f037184f2278c22100001013d2)