Ensure Android Studio errors handlers are run
This is a temp fix for 4.0 that will enable our 4.0 error handlers which
were disabled due to a platform change.

This change removes the isInternalCheck which was making some errors not
hit the error handlers, it also removes some parsers that were added to
the platform to ensure that errors aren't reported twice.

The handlers will still not be run when running the Android Plugin with
IDEA, but this should be fine.

Bug: 153558436
Test: N/A
Change-Id: I3d6981ad46d9dc84fe15d0d295ee70072b8b2141