Change loading of incomplete non-bundled plugins
When an incompatible non-bundled plugin is loaded, its plugin ID is
added to PluginLoadingResult.incompletePlugins, which causes plugins
that depend on it not to be loaded, even if later a compatible version
is loaded. This change adds incompatible non-bundled plugins to a
separate map, and merges the new map to incompletePlugins before
initializePlugins is called. Plugins depending on incomplete
dependencies are still caught during initializePlugins, since the
incomplete ones are not added to idMap.

Bug: 155919351
Test: existing plus manual tests with an incompatible plugin in
the local AndroidPreview directory

Change-Id: I79bb718fdbf1efb2fc5759dd86d641950d9e157f
(cherry picked from commit 3d99731b62ec70c409bca6672991621ba50eaf5a)