Remove references to plugin XML fragments that don't exist
The groovy and java plugin xmls reference some non existent
plugin xml fragments. These are supposed to be loaded only
in Ultimate edition. However, the presence of this results
in a a few harmless exceptions during the IDE startup.

This is a temporary fix (in 4.2 only) so that we don't pollute
the logs unnecessarily. Hopefully JB will do a proper fix
via IDEA-256190.

Bug: https://youtrack.jetbrains.com/issue/IDEA-256190
Test: launch IDE, observe no exceptions at startup
Change-Id: I9fe59f22256acc8c3032d7608ff2a7a9b76ec5bb