Fix Project leak through VcsToolWindowDnDTarget
IntelliJ commit 4f5f8e3c0f42d0d62b36bfdc724495c0e53be85a changed the
default parent disposable for SingleAlarm from null to
ApplicationManager.getApplication(). The runnable for the alarm in
VcsToolWindowDndTarget captures 'this', which contains a reference to
the project.

Instead, use the Content object as the parent disposable for the alarm.

Bug: 233632675
Change-Id: I100a51fbfc694cd0665c530d59d8245e4c867a58