Fix handling of not updated VFS in coverage plugin
The Coverage plugin used to expect build output directories to be
present in the virtual file system as a result of building a project.
However, even making sure that build output roots are present in the VFS
is an expensive operation on large projects. This is despite the
coverage plugin using Java IO to deal with .class files.

Use `getCompilerOutputUrl()` instead of `getCompilerOutput()` and make
sure the virtual file system is refreshed when it is really needed. This
avoids refreshing on build completion.

(cherry picked from commit daa5117a659028e27b262c80c09a5dc21b530df7)

Change-Id: If5e2c8075bfb1733e66e9922d90551714cd1090e