Dispose the MessageBus listener on dispose
The current code assumes that the frame is only disposed by receiving a
projectOpened event. However, Studio will create two WelcomeFrames in
some circumstances, calling dispose() on the first directly, then
creating the second.

When dispose() is called directly, the MessageBus listener is not
removed; this is a memory leak. It also causes a bug when the
projectOpened event occurs later: cleanup is performed by the dangling
listener first, causing the second frame's cleanup method to fail.

Bug: 219085149
Change-Id: I044ebcf58c1b3eb266750afd209e66addf43f54e