Allow GrMethodWrapper objects to be used as keys
Hashing data structures require their keys' hashCode() to be unchanging
while the object is used as a key.  GrMethodWrapper lazily computes its
return type, which means that the underlying myReturnType field in the
GrLightMethodBuilder which is used in the hashCode() (and equality)
can change.  Convert the direct use of the private field to a use of the
overridden accessor.

This had visible effects because GrMethodWrapper objects are used as
keys in the BoundedLocalCache used by the IconDeferrer implementation.

closes https://github.com/JetBrains/intellij-community/pull/2556

GitOrigin-RevId: 059b2fe642d9d82da1fda38fcc92ebc9d1a9b4cf

(cherry picked from commit cc7f5e268a64ce59bfdcf1b06525799888ffcf11)

Change-Id: Icc7f5e268a64ce59bfdcf1b06525799888ffcf11
Fixes: 297231210