Remove the Lombok plugin from Android Studio
The Lombok plugin induces runtime exceptions during indexing, even for
projects that do not use Lombok. Note that Lombok is already unbundled
in both AS Giraffe and AS Iguana. Starting in AS Iguana, users can
download the Lombok plugin on-demand from the plugin marketplace.

This reverts Change I8c88e154f.

Fixes: 301457783
Bug: 266333624
Change-Id: Ibeb13a44975d1cc0d2bb1e96bb429eb8426ddec7